Renamed the 'train' package to 'iPEPS'
