+++ src/pepsflow/optimize.cfg
@@ -1,18 +1,18 @@
 [PARAMETERS]
 model = Heisenberg
 chi = 72
-D = [2]
+D = [3]
 lam = 1
 warmup_steps = 100
 line_search = True
 runs = 1
 learning_rate = 0.5
-epochs = 100
+epochs = 50
 perturbation = 0.0
 gpu = False
 threads = 1
 save_folder = test2
-data_folder = test2
+data_folder = None
 start_epoch = -1
 split = False
