+++ tests/test_iPEPS_trainer.py
@@ -1,6 +1,6 @@
 import pytest
 
-from pepsflow.train.iPEPS_trainer import iPEPSTrainer
+from pepsflow.iPEPS.trainer import iPEPSTrainer
 
 
 class TestiPEPSTrainer:
